requirements.txt
aiosqlite
python-dotenv
httpx
